Move scaffold action types into util/scaffold package
