+++ app/src/main/java/com/msmobile/visitas/settings/SettingsScreen.kt
@@ -41,9 +41,9 @@ import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.msmobile.visitas.AppScaffold
-import com.msmobile.visitas.AppScaffoldState
+import com.msmobile.visitas.util.scaffold.AppScaffoldState
 import com.msmobile.visitas.R
-import com.msmobile.visitas.topNavigationActions
+import com.msmobile.visitas.util.scaffold.topNavigationActions
 import com.msmobile.visitas.extension.showShareIntent
 import com.msmobile.visitas.ui.theme.PreviewFoldable
 import com.msmobile.visitas.ui.theme.PreviewPhone
