+++ app/src/main/java/com/msmobile/visitas/util/scaffold/TopBarAction.kt
@@ -1,4 +1,4 @@
-package com.msmobile.visitas
+package com.msmobile.visitas.util.scaffold
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.vector.ImageVector
